voice-agent-backend/requirements.txt
livekit-agents==0.7.1
livekit==0.17.4
python-dotenv==1.0.1
httpx==0.27.2
openai==0.28.1
anthropic==0.7.8
python-dateutil==2.8.2
pydantic==2.6.4
tiktoken==0.5.2
uvicorn==0.27.1
fastapi==0.110.0
websockets==12.0
deepgram-sdk==3.1.2
cartesia==1.4.0
orjson==3.9.15
